streamlit_app/requirements.txt
streamlit
deepface
numpy
scipy
Pillow
tf-keras
opencv-python
